Adding reference histogram matching support (reference made using hist_matcher)

+++ rh_renderer/multiple_tiles_affine_renderer.py
@@ -104,9 +104,9 @@ class MultipleTilesAffineRenderer:
                 t_img, t_start_point, t_weights = t.crop_with_distances(from_x, from_y, to_x, to_y)
                 if t_img is not None:
                     res[t_start_point[1] - from_y: t_img.shape[0] + (t_start_point[1] - from_y),
-                        t_start_point[0] - from_x: t_img.shape[1] + (t_start_point[0] - from_x)] += t_img * t_weights
+                        t_start_point[0] - from_x: t_img.shape[1] + (t_start_point[0] - from_x)] += (t_img * t_weights).astype(np.uint32)
                     res_weights[t_start_point[1] - from_y: t_img.shape[0] + (t_start_point[1] - from_y),
-                                t_start_point[0] - from_x: t_img.shape[1] + (t_start_point[0] - from_x)] += t_weights
+                                t_start_point[0] - from_x: t_img.shape[1] + (t_start_point[0] - from_x)] += t_weights.astype(np.uint16)
 
             # Change the weights that are 0 to 1, to avoid division by 0
             res_weights[res_weights < 1] = 1
